Cherry pick to 0.15.x: Make tf.browser.fromPixels and route tf.fromPixels to it.
