+++ src/index.ts
@@ -27,6 +27,7 @@ import {deprecationWarn, disableDeprecationWarnings, enableProdMode, Environment
 // Serialization.
 import * as io from './io/io';
 import * as math from './math';
+import * as browser from './ops/browser';
 import * as serialization from './serialization';
 import {setOpHandler} from './tensor';
 import * as test_util from './test_util';
@@ -70,7 +71,7 @@ export {version as version_core};
 export {nextFrame, enableProdMode, disableDeprecationWarnings, deprecationWarn};
 
 // Second level exports.
-export {environment, io, math, serialization, test_util, util, webgl};
+export {browser, environment, io, math, serialization, test_util, util, webgl};
 
 // Backend specific.
 export {KernelBackend, BackendTimingInfo, DataMover, DataStorage} from './kernels/backend';
